feat(assertion-block): supply assertion error message
This makes assertion errors more descriptive
in case power-assert is not used.
It also happens to resolve the problem of
different assertion error messages
in different Node versions.

BREAKING CHANGE:
Generated assertions may now contain a message
as a second parameter to the assert function.
AssertionPostProcessors will need to take this into account
both for returning a power-assert pattern
and for their own AST transformations.